civicpulse/src/processing/requirements.txt
pymupdf>=1.23.0
pillow>=10.0.0
pytesseract>=0.3.10

